refactor(engine): decompose hard_links.rs into submodules (SPL-24 #2452)
Splits the 1325-line hard_links.rs into a focused submodule tree
under hard_links/: cohort (HardlinkApplyTracker leader/follower),
tracker (cross-platform facade), unix/windows (platform-specific
HardLinkTracker), and a tests/ subdirectory grouping the existing
test modules by concern. Zero API change; all re-exports preserved.
