+++ crates/engine/src/local_copy/hard_links/tests/mod.rs
@@ -1,0 +1,16 @@
+//! Test modules for the `hard_links` decomposition.
+
+mod apply_tracker_tests;
+mod io_uring_linkat_dispatch_tests;
+mod tracker_tests;
+
+#[cfg(unix)]
+mod cross_device_tests;
+#[cfg(unix)]
+mod detection_tests;
+#[cfg(unix)]
+mod device_inode_tests;
+#[cfg(unix)]
+mod preservation_tests;
+#[cfg(unix)]
+mod scale_tests;
